refactor: migrate source files to core subdirectory and update module imports
