+++ src/core/commands.ts
@@ -1,10 +1,10 @@
 import { Editor, TFile } from "obsidian";
-import { deleteXoppAndPdf, findCorrespondingXoppToPdf, openXournalppFile, renameXoppFile } from "./xoppActions";
-import XoppPlugin from "src/main";
-import { exportAllXoppToPDF, exportXoppToPDF } from "./xopp2pdf";
-import CreateXoppModalManager from "./CreateXoppModalManager";
-import RenameModal from "./modals/RenameModal";
-import SearchXoppModal from "./modals/SearchXoppModal";
+import { deleteXoppAndPdf, findCorrespondingXoppToPdf, openXournalppFile, renameXoppFile } from "../xoppActions";
+import XoppPlugin from "../main";
+import { exportAllXoppToPDF, exportXoppToPDF } from "../xopp2pdf";
+import CreateXoppModalManager from "../CreateXoppModalManager";
+import RenameModal from "../modals/RenameModal";
+import SearchXoppModal from "../modals/SearchXoppModal";
 
 export function createCommands(plugin: XoppPlugin) {
     plugin.addCommand({
