+++ src/core/default-template.ts
