+++ src/core/environment-checks.ts
